requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-ai
python-dotenv